rename aug file
