+++ molclr_torch_amp.py
@@ -193,8 +193,8 @@ def main():
         from dataset.dataset import MoleculeDatasetWrapper
     elif config["aug"] == "subgraph":
         from dataset.dataset_subgraph import MoleculeDatasetWrapper
-    elif config['aug'] == 'abbre':
-        from dataset.dataset_abbre import MoleculeDatasetWrapper
+    elif config['aug'] == 'func':
+        from dataset.dataset_func import MoleculeDatasetWrapper
     elif config["aug"] == "mix":
         from dataset.dataset_mix import MoleculeDatasetWrapper
     elif config['aug'] == 'no_aug':
